fix: created hooks for app
